NOT WORKING ANG PAG MAKE NG NOTES
