+++ app/src/androidTest/java/com/example/journalapp/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.notesapptutorial;
+package com.example.journalapp;
 
 import android.content.Context;
