+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#f9f9f9"
-    tools:context=".MainActivity">
+    tools:context="com.example.journalapp.MainActivity">
 
 
     <ProgressBar
