+++ app/src/main/res/layout/activity_notedetails.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".notedetails">
+    tools:context="com.example.journalapp.notedetails">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbarofnotedetail"
