+++ app/src/main/res/layout/activity_notesactivity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".notesactivity">
+    tools:context="com.example.journalapp.notesactivity">
 
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
