+++ app/src/main/res/layout/activity_signup.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#f9f9f9"
-    tools:context=".signup">
+    tools:context="com.example.journalapp.signup">
