+++ build.gradle
@@ -3,10 +3,11 @@ buildscript {
     repositories {
         google()
         jcenter()
+
     }
     dependencies {
         classpath "com.android.tools.build:gradle:4.1.1"
-        classpath 'com.google.gms:google-services:4.3.4'
+        classpath 'com.google.gms:google-services:4.3.15'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -17,6 +18,7 @@ allprojects {
     repositories {
         google()
         jcenter()
+
     }
 }
